fix some 64bit warnings
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2543 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ codeview.h
@@ -276,7 +276,7 @@ public:
 	bool DisplayAutoComplete;
 	void GetMembers(vector<UserData>* ListIn, const string &StrIn);
 	//TODO: int TabStop;
-	size_t DisplayAutoCompleteLength;
+	int DisplayAutoCompleteLength;
 	bool DwellDisplay;
 	bool DwellHelp;
 	int DwellDisplayTime;
